refactor: extract HTTP routing and handlers into dedicated router files.
